Test to_jsonapi and slightly reorganise actual test utilities
