+++ src/minerva_db/sql/api/utils.py
@@ -1,4 +1,4 @@
-def to_jsonapi(data, included=[]):
+def to_jsonapi(data, included={}):
     return {
         'data': data,
         'included': included
