+++ tests/sql/subjects.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm.exc import NoResultFound
 from minerva_db.sql.models import User, Group, Membership
 from minerva_db.sql.api.utils import to_jsonapi
 from .factories import GroupFactory, UserFactory, MembershipFactory
-from .utils import sa_obj_to_dict
+from . import sa_obj_to_dict
 
 
 class TestUser():
